% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{btlg}
\alias{btlg}
\title{A helper function returning stable beta, tau, and log(1+a*exp(g)).}
\usage{
btlg(g, a, what = c("b", "tau"))
}
\arguments{
\item{what}{}
}
\value{
A list containing:
ind -- indices of y_i for which g_i is very small.
ii -- indices of y_i for which g_i is very large.
}
\description{
A helper function returning stable beta, tau, and log(1+a*exp(g)).
}
