% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB-package.R
\docType{package}
\name{hurdleNB-package}
\alias{hurdleNB-package}
\title{hurdleNB: Hurdle negative binomial family for mgcv}
\description{
An implementation of the hurdle negative binomial extended GAM family for use with mgcv.
}
\author{
\strong{Maintainer}: Victor Tuekam \email{tuekam@lmu.de} (\href{https://orcid.org/0000-0002-0973-3492}{ORCID})

}
\keyword{internal}
