% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB.R
\name{hurdleNB}
\alias{hurdleNB}
\title{Hurdle negative binomial extended family for use with mgcv's \code{gam} or
\code{bam}.}
\usage{
hurdleNB(theta = NULL, link = "identity", b = 0)
}
\arguments{
\item{theta}{\eqn{\vartheta}, a numeric vector containing the 3 parameters
of the model, \eqn{\vartheta_0}, \eqn{\vartheta_1},
\eqn{\vartheta_2},}

\item{link}{Link function name, a character string or function name.
Only the identity -- \eqn{\log}-\eqn{\log} link and the
\eqn{\log} link for \eqn{q} and \eqn{\mu}, respectively,
are currently supported,}

\item{b}{A non-negative numeric parameter, specifying the minimum
dependence of the zero-inflation rate on the linear predictor.}
}
\value{
An object of class \code{extended.family}.
}
\description{
Hurdle negative binomial extended family for use with mgcv's \code{gam} or
\code{bam}.
}
\details{
\eqn{(\vartheta_1, b + e^{\vartheta_2}, \vartheta_0)} are the \code{theta}
parameters reported in the model summary.
}
\examples{
library(mgcv)
library(hurdleNB)

set.seed(1)
n <- 400
dat <- gamSim(1, n = n)
dat$y <- rhurdleNB(dat$f / 4 - 1)

m <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3),
  data = dat, family = hurdleNB()
)

m$outer.info # check convergence!
plot(m, page = 1)
plot(m, pages = 1, unconditional = TRUE)
}
