% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB_ll.R
\name{hurdleNB_ll}
\alias{hurdleNB_ll}
\title{Evaluate hurdle negative binomial log-likelihood and its derivatives.}
\usage{
hurdleNB_ll(y, g, eta, th0, level = 0)
}
\arguments{
\item{y}{\eqn{y}, a numeric vector,}

\item{g}{\eqn{\gamma}, a numeric vector,}

\item{eta}{\eqn{\eta}, a numeric vector,}

\item{th0}{\eqn{\vartheta_0}, a numeric,}

\item{level}{\itemize{
\item \eqn{== 0} - eval,
\item \eqn{> 0}  - derivatives for estimating \eqn{\beta}
and \eqn{\rho} using quasi-Newton,
\item \eqn{> 1}  - derivatives for estimating \eqn{\beta}
and \eqn{\rho} using full Newton.
}}
}
\value{
hurdleNB log-likelihood and its derivatives.
}
\description{
Evaluate hurdle negative binomial (NB) log-likelihood and its derivatives
w.r.t. \eqn{\gamma} (g) and \eqn{\eta} (eta), with
\eqn{1 - q = e^{-e^{\eta}}} and \eqn{\mu = e^{\gamma}}, for each datum in
vector \eqn{y}. \eqn{q} is the probability of potential presence.
\eqn{\mu} is the NB mean.
}
