% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB_ll.R
\name{ktlg}
\alias{ktlg}
\title{A helper function returning stable
\eqn{\kappa}, \eqn{\tau}, and \eqn{\log(1 + \alpha e^{\gamma})}.}
\usage{
ktlg(g, a, what = c("k", "tau"))
}
\arguments{
\item{g}{\eqn{\gamma}, a numeric vector,}

\item{a}{\eqn{\alpha}, a numeric,}

\item{what}{A character vector specifying what to return.}
}
\value{
A list containing:
\itemize{
\item \code{k}   – \eqn{\kappa},
\item \code{tau} – \eqn{\tau},
\item \code{lg}  – \eqn{\log(1 + \alpha e^{\gamma})},
\item \code{ind} – indices of \eqn{y_i} for which \eqn{\gamma_i}
is very small,
\item \code{ii}  – indices of \eqn{y_i} for which \eqn{\gamma_i}
is very large.
}
}
\description{
A helper function returning stable
\eqn{\kappa}, \eqn{\tau}, and \eqn{\log(1 + \alpha e^{\gamma})}.
}
