% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{l1ee}
\alias{l1ee}
\title{log(1-exp(-exp(ğ›„))).}
\usage{
l1ee(g)
}
\arguments{
\item{g}{\itemize{
\item ğ›„, a numeric vector,
}}
}
\value{
Carefully computed log(1-exp(-exp(ğ›„))).
}
\description{
log(1-exp(-exp(ğ›„))).
}
