% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB_ll.R
\name{ldg}
\alias{ldg}
\title{Log-likelihood derivatives w.r.t. ğ›„.}
\usage{
ldg(g, y, a, v, level = 2)
}
\arguments{
\item{g}{\itemize{
\item ğ›„, a numeric vector,
}}

\item{y}{\itemize{
\item ğ², a numeric vector,
}}

\item{a}{\itemize{
\item Î±, a numeric,
}}

\item{v}{\itemize{
\item v, a list containing ğ›‹ and ğ›•,
}}

\item{level}{\itemize{
\item == 0 - first and second derivatives,
>  0 - derivatives needed for quasi-Newton,
>  1 - derivatives need for full Newton.
}}
}
\value{
A list of derivatives of the log-likelihood w.r.t. ğ›„ (g).
}
\description{
Log-likelihood derivatives w.r.t. ğ›„.
}
