% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{ldgth0}
\alias{ldgth0}
\title{Log-likelihood (â„“) derivatives w.r.t. Î¸â‚€.
and mixed derivatives of â„“ w.r.t. ğ›„ and Î¸â‚€.}
\usage{
ldgth0(g, y, th0, v, level = 2)
}
\arguments{
\item{g}{\itemize{
\item ğ›„, a numeric vector,
}}

\item{y}{\itemize{
\item ğ², a numeric vector,
}}

\item{th0}{\itemize{
\item Î¸â‚€, a numeric,
}}

\item{v}{\itemize{
\item v, a list containing ğ›‹, ğ›• and lg,
}}

\item{level}{\itemize{
\item == 0 - list of NULLs (not needed for estimating parameters),
>  0 - derivatives needed for quasi-Newton,
>  1 - derivatives need for full Newton.
}}
}
\value{
A list of the first, second amd mixed derivatives of the
log-likelihood w.r.t. Î¸â‚€ and ğ›„.
}
\description{
Log-likelihood (â„“) derivatives w.r.t. Î¸â‚€.
and mixed derivatives of â„“ w.r.t. ğ›„ and Î¸â‚€.
}
