% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{ldth0}
\alias{ldth0}
\title{Log-likelihood derivatives w.r.t. a.}
\usage{
ldth0(g, y, th0)
}
\arguments{
\item{a}{}
}
\value{
A list of the first and second derivatives of the
}
\description{
Log-likelihood derivatives w.r.t. a.
}
