% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB.R
\name{lind}
\alias{lind}
\title{\eqn{\eta = \vartheta_1 + (b + e^{\vartheta_2})\gamma}.}
\usage{
lind(g, theta, level = 0, b = 0)
}
\arguments{
\item{g}{\eqn{\gamma}, a numeric vector,}

\item{theta}{\eqn{\vartheta}, a numeric vector,}

\item{level}{A numeric, indicating whether to return deriv
w.r.t. \eqn{\vartheta_1} and \eqn{\vartheta_2},}

\item{b}{A numeric.}
}
\value{
A list with \eqn{\eta = \vartheta_1 + (b + e^{\vartheta_2})\gamma}
and its derivatives with respect to \eqn{\gamma}, \eqn{\vartheta_1},
and \eqn{\vartheta_2}.
}
\description{
\eqn{\eta = \vartheta_1 + (b + e^{\vartheta_2})\gamma}.
}
