% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB.R
\name{rhurdleNB}
\alias{rhurdleNB}
\title{Generate hurdle NB random variables.}
\usage{
rhurdleNB(g, theta = c(-2, 0.3, 2), b = 0)
}
\arguments{
\item{g}{\itemize{
\item \eqn{\gamma}, a numeric vector,
}}

\item{theta}{\itemize{
\item \eqn{\vartheta}, a numeric vector,
}}

\item{b}{\itemize{
\item A numeric.
}}
}
\value{
hurdle negative binomial random variables.
}
\description{
Generate hurdle NB random variables.
}
