% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb.R
\name{rzinb}
\alias{rzinb}
\title{Generate zero-inflated NB random variables.}
\usage{
rzinb(g, theta = c(-2, 0.3, 2), b = 0)
}
\arguments{
\item{g}{\itemize{
\item ğ², a numeric vector,
}}

\item{theta}{\itemize{
\item ğ›‰, a numeric vector,
}}

\item{b}{\itemize{
\item A numeric.
}}
}
\value{
zero-inflated negative binomial random variables.
}
\description{
Generate zero-inflated NB random variables.
}
