% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb-package.R
\docType{package}
\name{zinb-package}
\alias{zinb-package}
\title{zinb: Zero-inflated negative binomial family for mgcv}
\description{
An implementation of the zero-inflated negative binomial extended GAM family for use with mgcv.
}
\author{
\strong{Maintainer}: Victor Tuekam \email{tuekam@ifo.de} (\href{https://orcid.org/0000-0002-0973-3492}{ORCID})

}
\keyword{internal}
