% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb.R
\name{zinb}
\alias{zinb}
\title{Zero-inflated negative binomial extended family for mgcv.}
\usage{
zinb(theta = NULL, link = "identity", b = 0)
}
\arguments{
\item{theta}{\itemize{
\item ğ›‰, a numeric vector containing the 3 parameters of the model,
Î¸â‚€, Î¸â‚, Î¸â‚‚,
}}

\item{link}{\itemize{
\item link function name, a character string or function name,
}}

\item{b}{\itemize{
\item a numeric parameter.
}}
}
\value{
An object of class \code{extended.family}.
}
\description{
Zero-inflated negative binomial extended family for mgcv.
}
\examples{
library(mgcv)
set.seed(1)
n <- 400
dat <- gamSim(1, n = n)
dat$y <- rzinb(dat$f / 4 - 1)
m <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3),
  family = zinb(), data = dat,
  optimizer = c("outer", "bfgs")
)
m$outer.info # check convergence!
plot(m, page = 1)
}
