% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{zinbll}
\alias{zinbll}
\title{Evaluate zero-inflated negative binomial log-likelihood
and its derivatives w.r.t. ğ›„ (g) and ğ›ˆ (eta), with
1-q = exp(-exp(ğ›ˆ)) and ğ› = exp(ğ›„), for each datum in vector y.
q is probability of potential presence. mu is the NB mean.}
\usage{
zinbll(y, g, eta, th0, deriv = 0)
}
\arguments{
\item{y}{\itemize{
\item ğ², a numeric vector
}}

\item{g}{\itemize{
\item ğ›„, a numeric vector
}}

\item{eta}{\itemize{
\item ğ›ˆ, a numeric vector
}}

\item{th0}{\itemize{
\item Î¸â‚€, a numeric
}}

\item{deriv}{\itemize{
\item 0 - eval.
1 - gradient and Hessian.
2 - third derivatives.
4 - fourth derivatives.
}}
}
\value{
ZINB log-likelihood and its derivatives.
}
\description{
Evaluate zero-inflated negative binomial log-likelihood
and its derivatives w.r.t. ğ›„ (g) and ğ›ˆ (eta), with
1-q = exp(-exp(ğ›ˆ)) and ğ› = exp(ğ›„), for each datum in vector y.
q is probability of potential presence. mu is the NB mean.
}
