% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{btlg}
\alias{btlg}
\title{A helper function returning stable ğ›ƒ, ğƒ, and log(1+Î±eáµ).}
\usage{
btlg(g, a, what = c("b", "tau"))
}
\arguments{
\item{g}{\itemize{
\item ğ›„, a numeric vector.
}}

\item{a}{\itemize{
\item Î±, a numeric.
}}

\item{what}{\itemize{
\item A character vector specifying what to return.
}}
}
\value{
A list containing:
ind -- indices of yáµ¢ for which Î³áµ¢ is very small.
ii -- indices of yáµ¢ for which Î³áµ¢ is very large.
}
\description{
A helper function returning stable ğ›ƒ, ğƒ, and log(1+Î±eáµ).
}
