% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB_ll.R
\name{hurdleNB_ll}
\alias{hurdleNB_ll}
\title{Evaluate hurdle negative binomial log-likelihood
and its derivatives w.r.t. ğ›„ (g) and ğ›ˆ (eta), with
1-q = exp(-exp(ğ›ˆ)) and ğ› = exp(ğ›„), for each datum in vector y.
q is probability of potential presence. mu is the NB mean.}
\usage{
hurdleNB_ll(y, g, eta, th0, level = 0)
}
\arguments{
\item{y}{\itemize{
\item ğ², a numeric vector,
}}

\item{g}{\itemize{
\item ğ›„, a numeric vector,
}}

\item{eta}{\itemize{
\item ğ›ˆ, a numeric vector,
}}

\item{th0}{\itemize{
\item Î¸â‚€, a numeric,
}}

\item{level}{\itemize{
\item == 0 - eval,
>  0 - derivatives for estimating ğ›ƒ and ğ›’ using quasi-Newton,
>  1 - derivatives for estimating ğ›ƒ and ğ›’ using full Newton.
}}
}
\value{
hurdleNB log-likelihood and its derivatives.
}
\description{
Evaluate hurdle negative binomial log-likelihood
and its derivatives w.r.t. ğ›„ (g) and ğ›ˆ (eta), with
1-q = exp(-exp(ğ›ˆ)) and ğ› = exp(ğ›„), for each datum in vector y.
q is probability of potential presence. mu is the NB mean.
}
