% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{l11aea}
\alias{l11aea}
\title{log(1 - (1 + a*exp(x))^(-1/a)).}
\usage{
l11aea(x, th0)
}
\arguments{
\item{a}{}
}
\value{
Carfully compute log(1 - (1 + a*exp(x))^(-1/a)).
}
\description{
log(1 - (1 + a*exp(x))^(-1/a)).
}
