% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{l11aea}
\alias{l11aea}
\title{log((1 + Î±exp(ğ›„))^(1/Î±) - 1).}
\usage{
l11aea(g, th0)
}
\arguments{
\item{g}{\itemize{
\item ğ›„, a numeric vector,
}}

\item{th0}{\itemize{
\item Î¸â‚€, a numeric.
}}
}
\value{
Carfully compute log((1 + exp(Î¸â‚€)exp(ğ›„))^(1/(exp(Î¸â‚€))) - 1).
}
\description{
log((1 + Î±exp(ğ›„))^(1/Î±) - 1).
}
