% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB_ll.R
\name{l11aea}
\alias{l11aea}
\title{\eqn{\log((1 + \alpha e^{\gamma})^{\frac{1}{\alpha}} - 1)}}
\usage{
l11aea(g, th0)
}
\arguments{
\item{g}{\eqn{\gamma}, a numeric vector,}

\item{th0}{\eqn{\vartheta_0}, a numeric.}
}
\value{
Carefully computed
\eqn{\log((1 + e^{\vartheta_0} e^{\gamma})^{\frac{1}{e^{\vartheta_0}}} - 1)}.
}
\description{
\eqn{\log((1 + \alpha e^{\gamma})^{\frac{1}{\alpha}} - 1)}
}
