% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{l11aea}
\alias{l11aea}
\title{log(1 - (1 + αeˣ)^(-1/α)).}
\usage{
l11aea(x, th0)
}
\arguments{
\item{x}{\itemize{
\item A numeric vector.
}}

\item{th0}{\itemize{
\item θ₀, a numeric.
}}
}
\value{
Carfully compute log(1 - (1 + exp(θ₀)exp(x))^(-1/(exp(θ₀)))).
}
\description{
log(1 - (1 + αeˣ)^(-1/α)).
}
