% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB_ll.R
\name{l1ee}
\alias{l1ee}
\title{\eqn{\log(1 - e^{-e^{\eta}})}.}
\usage{
l1ee(eta)
}
\arguments{
\item{eta}{\itemize{
\item \eqn{\eta}, a numeric vector.
}}
}
\value{
Carefully computed \eqn{\log(1 - e^{-e^{\eta}})}.
}
\description{
\eqn{\log(1 - e^{-e^{\eta}})}.
}
