% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB_ll.R
\name{l1ee}
\alias{l1ee}
\title{log(1-exp(-exp(ğ›ˆ))).}
\usage{
l1ee(eta)
}
\arguments{
\item{eta}{\itemize{
\item ğ›ˆ, a numeric vector.
}}
}
\value{
Carefully computed log(1-exp(-exp(ğ›ˆ))).
}
\description{
log(1-exp(-exp(ğ›ˆ))).
}
