% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB_ll.R
\name{lde}
\alias{lde}
\title{Log-likelihood derivatives w.r.t. \eqn{\eta}.}
\usage{
lde(eta, level = 2)
}
\arguments{
\item{eta}{\eqn{\eta}, a numeric vector,}

\item{level}{\itemize{
\item \eqn{==0} – first and second derivatives,
\item \eqn{> 0} – derivatives needed for quasi-Newton,
\item \eqn{> 1} – derivatives needed for full Newton.
}}
}
\value{
A list of derivatives of the log-likelihood w.r.t. \eqn{\eta} (eta).
}
\description{
Log-likelihood derivatives w.r.t. \eqn{\eta}.
}
