% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleNB_ll.R
\name{ldg}
\alias{ldg}
\title{Log-likelihood derivatives w.r.t. \eqn{\gamma}.}
\usage{
ldg(g, y, a, v, level = 2)
}
\arguments{
\item{g}{\itemize{
\item \eqn{\gamma}, a numeric vector,
}}

\item{y}{\itemize{
\item \eqn{y}, a numeric vector,
}}

\item{a}{\itemize{
\item \eqn{\alpha}, a numeric,
}}

\item{v}{\itemize{
\item \code{v}, a list containing \eqn{\kappa}, \eqn{\tau}, \code{ind} and \code{ii},
}}

\item{level}{\itemize{
\item \eqn{0} – first and second derivatives.
\item \eqn{> 0} – derivatives needed for quasi-Newton.
\item \eqn{> 1} – derivatives needed for full Newton.
}}
}
\value{
A list of derivatives of the log-likelihood w.r.t. \eqn{\gamma} (g).
}
\description{
Log-likelihood derivatives w.r.t. \eqn{\gamma}.
}
