% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{ldg}
\alias{ldg}
\title{Log-likelihood derivates w.r.t. g.}
\usage{
ldg(g, y, a, deriv = 4)
}
\arguments{
\item{deriv:}{<= 1 - first and second derivatives.
== 2 - first, second and third derivatives.
>= 3 - first, second, third and fourth derivatives.}
}
\value{
A list of derivatives of the log-likelihood w.r.t. g.
}
\description{
Log-likelihood derivates w.r.t. g.
}
