% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{ldgth0}
\alias{ldgth0}
\title{Mixed derivatives of l w.r.t. ğ›„ and Î¸â‚€.}
\usage{
ldgth0(g, y, th0, v, deriv = 4)
}
\arguments{
\item{g}{\itemize{
\item ğ›„, a numeric vector,
}}

\item{y}{\itemize{
\item ğ², a numeric vector,
}}

\item{th0}{\itemize{
\item Î¸â‚€, a numeric,
}}

\item{v}{\itemize{
\item v, a list containing ğ›‹, ğ›• and lg,
}}

\item{deriv}{\itemize{
\item <= 1 - second mixed derivatives,
== 2 - second and third mixed derivatives,
>= 3 - second, third and fourth mixed derivatives.
}}
}
\value{
A list of mixed derivatives of l w.r.t. ğ›„ and Î¸â‚€.
}
\description{
Mixed derivatives of l w.r.t. ğ›„ and Î¸â‚€.
}
