% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{ldth0}
\alias{ldth0}
\title{Log-likelihood derivatives w.r.t. Î¸â‚€.}
\usage{
ldth0(g, y, th0, v)
}
\arguments{
\item{g}{\itemize{
\item ğ›„, a numeric vector,
}}

\item{y}{\itemize{
\item ğ², a numeric vector,
}}

\item{th0}{\itemize{
\item Î¸â‚€, a numeric,
}}

\item{v}{\itemize{
\item v, a list containing ğ›‹, ğ›• and lg.
}}
}
\value{
A list of the first and second derivatives of the
log-likelihood w.r.t. Î¸â‚€.
}
\description{
Log-likelihood derivatives w.r.t. Î¸â‚€.
}
