% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb.R
\name{lind}
\alias{lind}
\title{ğ›ˆ = Î¸â‚ + (b + exp(Î¸â‚‚))ğ².}
\usage{
lind(g, theta, deriv = 0, b = 0)
}
\arguments{
\item{g}{\itemize{
\item ğ², a numeric vector.
}}

\item{theta}{\itemize{
\item ğ›‰, a numeric vector.
}}

\item{deriv}{\itemize{
\item A numeric, indicating whether to return deriv w.r.t. Î¸â‚ & Î¸â‚‚.
}}

\item{b}{\itemize{
\item A numeric.
}}
}
\value{
A list with ğ›ˆ = Î¸â‚ + (b + exp(Î¸â‚‚))ğ² and its derivatives
w.r.t. ğ², Î¸â‚ & Î¸â‚‚.
}
\description{
ğ›ˆ = Î¸â‚ + (b + exp(Î¸â‚‚))ğ².
}
