% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbll.R
\name{zinbll}
\alias{zinbll}
\title{Evaluate zero-inflated NB log-likelihood
and its derivatives w.r.t. g (gamma) and eta, with
1-q = exp(-exp(eta)) and mu = exp(g), for each datum in vector y.
p is probability of potential presence. mu is the NB mean.}
\usage{
zinbll(y, g, eta, th0, deriv = 0)
}
\arguments{
\item{y}{\itemize{
\item ğ², a numeric vector
}}

\item{g}{\itemize{
\item ğ›„, a numeric vector
}}

\item{eta}{\itemize{
\item ğ›ˆ, a numeric vector
}}

\item{th0}{\itemize{
\item Î¸â‚€, a numeric
}}

\item{deriv}{\itemize{
\item 0 - eval.
1 - gradient and Hessian.
2 - third derivatives.
4 - fourth derivatives.
}}
}
\value{
ZINB log-likelihood and its derivatives.
}
\description{
Evaluate zero-inflated NB log-likelihood
and its derivatives w.r.t. g (gamma) and eta, with
1-q = exp(-exp(eta)) and mu = exp(g), for each datum in vector y.
p is probability of potential presence. mu is the NB mean.
}
